Modified to include float numbers

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-user_income = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your total monthly expenses: "))
+user_income = float(input("Enter your monthly income: "))
+monthly_expenses = float(input("Enter your total monthly expenses: "))
 
 monthly_savings = user_income - monthly_expenses
